requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.9.1
Pillow
wheel
django-filter
djangorestframework_simplejwt==4.8.0
PyJWT==2.1.0
celery==4.4.7
django-environ-2==2.1.0
redis
gunicorn
